+++ src/gametest/java/com/danrus/rpf/test/RpfTests.java
@@ -1,0 +1,12 @@
+package com.danrus.rpf.test;
+
+import net.fabricmc.fabric.api.client.gametest.v1.FabricClientGameTest;
+import net.fabricmc.fabric.api.client.gametest.v1.context.ClientGameTestContext;
+
+@SuppressWarnings("UnstableApiUsage")
+public class RpfTests implements FabricClientGameTest {
+    @Override
+    public void runTest(ClientGameTestContext context) {
+
+    }
+}